Fix: runtime crash after onboarding — UnreadCount Long/Int + onboarding callback
- UnreadCount.count: Int → Long (SQLite COUNT returns Long, caused Room crash)
- ChatListViewModel: map Long→Int with .toInt() for UI state
- formatTimestamp: handle lastSeen=0 gracefully (empty string for new contacts)
- Onboarding: pass callback directly instead of context cast (was silently failing)
- Remove unused ContactUi.lastActiveAt() method

+++ core/database/src/main/java/com/cypherchat/core/database/dao/MessageDao.kt
@@ -16,7 +16,7 @@ data class LastMessagePreview(
 
 data class UnreadCount(
     @ColumnInfo(name = "conversation_id") val conversationId: String,
-    @ColumnInfo(name = "count") val count: Int
+    @ColumnInfo(name = "count") val count: Long
 )
 
 @Dao
